llm/pyproject.toml
[tool.poetry]
name = "ai-chatbot-llm-api"
version = "0.1.0"
description = ""
authors = ["Rishil Ramesh <rishil.ramesh@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.15"
langchain-google-genai = "^1.0.10"
python-dotenv = "^1.0.1"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
